+++ noxfiles/docker_nox.py
@@ -2,6 +2,7 @@
 import nox
 from constants_nox import (
     IMAGE,
+    IMAGE_DEV,
     IMAGE_LATEST,
     IMAGE_LOCAL,
     IMAGE_LOCAL_UI,
@@ -49,8 +50,23 @@ def build(session: nox.Session, image: str) -> None:
 
 
 @nox.session()
-def push(session: nox.Session) -> None:
+@nox.parametrize(
+    "tag",
+    [
+        nox.param("prod", id="prod"),
+        nox.param("dev", id="dev"),
+    ],
+)
+def push(session: nox.Session, tag: str) -> None:
     """Push the fidesctl Docker image to Dockerhub."""
-    session.run("docker", "tag", get_current_image(), IMAGE_LATEST, external=True)
-    session.run("docker", "push", IMAGE, external=True)
-    session.run("docker", "push", IMAGE_LATEST, external=True)
+
+    tag_matrix = {"prod": IMAGE_LATEST, "dev": IMAGE_DEV}
+
+    # Push either "ethyca/fidesctl:dev" or "ethyca/fidesctl:latest"
+    session.run("docker", "tag", get_current_image(), tag_matrix[tag], external=True)
+    session.run("docker", "push", tag_matrix[tag], external=True)
+
+    # Only push the tagged version if its for prod
+    # Example: "ethyca/fidesctl:1.7.0"
+    if tag == "prod":
+        session.run("docker", "push", IMAGE, external=True)